Conflicts:
	Classes/Controllers/PBWebHistoryController.m
	Classes/git/PBGitCommit.h
	Classes/git/PBGitCommit.m
	Classes/git/PBGitGrapher.mm
	Classes/git/PBGitHistoryList.m
	Classes/git/PBGitRevList.mm
	Classes/git/PBGitSHA.h
	Classes/git/PBGitSHA.m

+++ Classes/Views/GLFileView.m
@@ -6,9 +6,14 @@
 //  Copyright 2010 __MyCompanyName__. All rights reserved.
 //
 
-#import "GLFileView.h"
-#import "PBGitGradientBarView.h"
 #import <MGScopeBar/MGScopeBar.h>
+
+#import "GLFileView.h"
+#import "PBGitTree.h"
+#import "PBGitSHA.h"
+#import "PBGitCommit.h"
+#import "PBGitHistoryController.h"
+
 
 #define GROUP_LABEL				@"Label"			// string
 #define GROUP_SEPARATOR			@"HasSeparator"		// BOOL as NSNumber
@@ -110,25 +115,25 @@
 
 #pragma mark MGScopeBarDelegate methods
 
-- (int)numberOfGroupsInScopeBar:(MGScopeBar *)theScopeBar
+- (NSInteger)numberOfGroupsInScopeBar:(MGScopeBar *)theScopeBar
 {
 	return [self.groups count];
 }
 
 
-- (NSArray *)scopeBar:(MGScopeBar *)theScopeBar itemIdentifiersForGroup:(int)groupNumber
+- (NSArray *)scopeBar:(MGScopeBar *)theScopeBar itemIdentifiersForGroup:(NSInteger)groupNumber
 {
 	return [[self.groups objectAtIndex:groupNumber] valueForKeyPath:[NSString stringWithFormat:@"%@.%@", GROUP_ITEMS, ITEM_IDENTIFIER]];
 }
 
 
-- (NSString *)scopeBar:(MGScopeBar *)theScopeBar labelForGroup:(int)groupNumber
+- (NSString *)scopeBar:(MGScopeBar *)theScopeBar labelForGroup:(NSInteger)groupNumber
 {
 	return [[self.groups objectAtIndex:groupNumber] objectForKey:GROUP_LABEL]; // might be nil, which is fine (nil means no label).
 }
 
 
-- (NSString *)scopeBar:(MGScopeBar *)theScopeBar titleOfItem:(NSString *)identifier inGroup:(int)groupNumber
+- (NSString *)scopeBar:(MGScopeBar *)theScopeBar titleOfItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 {
 	NSArray *items = [[self.groups objectAtIndex:groupNumber] objectForKey:GROUP_ITEMS];
 	if (items) {
@@ -143,12 +148,12 @@
 }
 
 
-- (MGScopeBarGroupSelectionMode)scopeBar:(MGScopeBar *)theScopeBar selectionModeForGroup:(int)groupNumber
+- (MGScopeBarGroupSelectionMode)scopeBar:(MGScopeBar *)theScopeBar selectionModeForGroup:(NSInteger)groupNumber
 {
 	return [[[self.groups objectAtIndex:groupNumber] objectForKey:GROUP_SELECTION_MODE] intValue];
 }
 
-- (void)scopeBar:(MGScopeBar *)theScopeBar selectedStateChanged:(BOOL)selected forItem:(NSString *)identifier inGroup:(int)groupNumber
+- (void)scopeBar:(MGScopeBar *)theScopeBar selectedStateChanged:(BOOL)selected forItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 {
 	startFile=identifier;
 	NSString *path = [NSString stringWithFormat:@"html/views/%@", identifier];
